requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==6.7
Flask==0.12.2
idna==2.10
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.1.1
numpy==1.19.5
opencv-python==4.5.1.48
pbkdf2==1.3
picamera==1.13
python-dotenv==0.14.0
requests==2.25.1
RPi.GPIO==0.7.0
smbus==1.1.post2
smbus2==0.4.1
urllib3==1.26.4
Werkzeug==0.12.2
wifi==0.3.8
